Update with project deps

+++ examples/lle_examples/scurve.jl
@@ -7,9 +7,9 @@ include("plots_lle.jl")
 
 function scurve_dane(N)
     θ = LinRange(-1.5f0 * π, 1.5f0 * π, N)
-    x = sin.(θ) .+ 0.1f0 .* randn(Float32, N)
-    y = LinRange(0f0, 5f0, N) .+ 0.01f0 .* randn(Float32, N) 
-    z = sign.(θ) .+ 0.1f0 .* randn(Float32, N)
+    x = sin.(θ) .+ 0.1f0 .* randn(Float64, N)
+    y = LinRange(0f0, 5f0, N) .+ 0.01f0 .* randn(Float64, N) 
+    z = sign.(θ) .+ 0.1f0 .* randn(Float64, N)
     X = [x y z]'
     return X, θ
 end
